split up the controller into pieces
