+++ backend/src/models/schema.ts
@@ -1,7 +1,5 @@
 import mongoose, { Schema, Model } from 'mongoose';
 import { Todo } from './todo';
-import { User } from './user.interface';
-import mongooseUniqueValidator from 'mongoose-unique-validator';
 
 export const todoSchema : Schema = new mongoose.Schema(
 {
@@ -14,20 +12,3 @@ export const todoSchema : Schema = new mongoose.Schema(
 
 export const todoModel : Model<Todo> = mongoose.model('Todos', todoSchema);
 
-export const userSchema : Schema = new mongoose.Schema(
-{
-	email:
-	{
-		type: String,
-		required: true,
-		unique : true
-	},
-	password:
-	{
-		type: String,
-		required: true
-	}
-});
-
-userSchema.plugin(mongooseUniqueValidator);
-export const userModel : Model<User> = mongoose.model('Users', userSchema);
